+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ClosedShell3D: is_face_inside, get_subtraction_valid_faces, valid_intersection_faces, point_belongs
 - ContourMixin: delete_shared_contour_section, reorder_contour_at_point, are_extremity_points_touching
 - BSplineFace3D: neutral_fiber
+- surfaces.Plane3D: linesegment_intersections
 
 ### Refactor
 - ClosedShell3D: point_belongs, get_non_intersecting_faces
@@ -127,6 +128,7 @@ to calculate intersections between a plane 3d and bsplinecurve3d.
 - utils.parametric: fix contour2d_healing
 - BSplineSurface3D: ban useless attr in serialization
 - BSplineCurve: simplify
+- SphericalSurface3D: contour3d_to_2d
 - WireMixin: to_wire_with_linesegments (use new methods, for 2D and 3D)
 - ArcEllipse2d: point_belongs, abscissa, init.
 - Face3D: face_inside - now considers inners_contours